refactor: split useLedVisualization into state, renderers, and loop
useLedVisualization.ts was a 476-line module mixing three layers: the
pure VU-meter state machine, the per-LED paint functions, and the
React hook driving the rAF loop. Split along those seams:

- vuMeterState.ts: the pure per-voice state machine (level smoothing,
  bar heights, peak hold/decay, VU<->ambient cross-fade) plus the
  VoiceVuState/CrossFadeState types. No DOM access; unit-testable in
  isolation.
- ledRenderers.ts: renderAmbientLed/renderVuLed and their option
  types, with the column lookup tables and the private
  brightness/ripple helpers they share.
- useLedVisualization.ts: just the hook - refs, the rAF animation
  loop, and the pointer/ripple API (~190 lines). The VU state
  machine is re-exported here so existing imports keep working.

Also drops the exported-but-unused VisualizationMode type.

No behaviour change: the existing led-icon suites (78 tests,
including the VU state-machine cases that import from
useLedVisualization) pass unmodified, and LedIconGrid /
useMiniLedAnimation consumers are untouched.

Second slice of the audit's mega-hook backlog.
