+++ app/renderer/components/led-icon/useLedVisualization.ts
@@ -1,59 +1,38 @@
 import { useCallback, useEffect, useRef } from "react";
 
 import type { Ripple } from "../dialogs/led-grid/ledMath";
+import type { CrossFadeState, VoiceVuState } from "./vuMeterState";
 
+import { readGlowColor, readVoiceGlowColor } from "../dialogs/led-grid/ledMath";
+import { hasActiveVoices } from "./audioLevels";
 import {
-  applyLedStyle,
-  computeBaseLfo,
-  computeProximityBoost,
-  computeRippleBoost,
-  readGlowColor,
-  readVoiceGlowColor,
-} from "../dialogs/led-grid/ledMath";
-import { getVoiceLevel, hasActiveVoices } from "./audioLevels";
-import {
-  ICON_BASE_MIN,
-  ICON_BASE_SCALE,
   ICON_COLS,
   ICON_LED_COUNT,
   ICON_MAX_RIPPLES,
-  ICON_PROXIMITY_INTENSITY,
-  ICON_PROXIMITY_RADIUS_SQ,
   ICON_RIPPLE_DURATION,
-  ICON_RIPPLE_SPEED,
-  ICON_RIPPLE_WIDTH,
-  ICON_ROW_LFO_CONFIGS,
-  ICON_ROWS,
   ICON_TIME_HOVER,
   ICON_TIME_IDLE,
 } from "./ledIconConstants";
+import { renderAmbientLed, renderVuLed } from "./ledRenderers";
 import {
-  VU_AMPLIFICATION,
-  VU_BORDER_BRIGHTNESS,
-  VU_BORDER_COLS,
-  VU_CROSSFADE_MS,
-  VU_DECAY_RATE,
-  VU_LIT_MAX_BRIGHTNESS,
-  VU_LIT_MIN_BRIGHTNESS,
-  VU_NOISE_FLOOR,
-  VU_PEAK_BRIGHTNESS,
-  VU_PEAK_DROP_INTERVAL,
-  VU_PEAK_HOLD_FRAMES,
-  VU_POWER_CURVE,
-  VU_SILENCE_THRESHOLD,
-  VU_UNLIT_BRIGHTNESS,
-  VU_VOICE_COLS,
-} from "./vuMeterConstants";
+  computeVoiceBar,
+  getVuState,
+  updateCrossFade,
+  updatePeakState,
+} from "./vuMeterState";
 
-export type VisualizationMode = "classic" | "screensaver";
+// Re-export the pure VU state machine for existing consumers/tests
+export type { CrossFadeState, VoiceVuState } from "./vuMeterState";
 
-// Per-voice VU state
-export interface VoiceVuState {
-  peakDropCounter: number;
-  peakHoldCounter: number;
-  peakRow: number; // -1 = no peak
-  smoothedLevel: number;
-}
+export {
+  computeVoiceBar,
+  computeVuLedBrightness,
+  decayPeak,
+  getVuState,
+  initVuState,
+  updateCrossFade,
+  updatePeakState,
+} from "./vuMeterState";
 
 interface UseLedVisualizationReturn {
   addRipple: (col: number, row: number) => void;
@@ -62,176 +41,11 @@ interface UseLedVisualizationReturn {
   setMousePosition: (col: number, row: number) => void;
 }
 
-// Build column-to-voice lookup table
-const colToVoice = new Map<number, number>();
-for (const [voice, cols] of Object.entries(VU_VOICE_COLS)) {
-  for (const col of cols) {
-    colToVoice.set(col, Number(voice));
-  }
-}
-
-const borderColSet = new Set(VU_BORDER_COLS);
-
-interface AmbientLedOptions {
-  col: number;
-  el: HTMLElement;
-  glowColor: string;
-  isHovered: boolean;
-  mouse: { col: number; row: number } | null;
-  rawTime: number;
-  ripples: Ripple[];
-  row: number;
-  time: number;
-}
-
-interface CrossFadeState {
-  ambientFade: number;
-  vuInactiveSince: null | number;
-  wasVuActive: boolean;
-}
-
-interface VuLedOptions {
-  ambientMix: number;
-  col: number;
-  defaultGlowColor: string;
-  el: HTMLElement;
-  rawTime: number;
-  ripples: Ripple[];
-  row: number;
-  time: number;
-  voiceBarHeights: Record<number, number>;
-  voiceGlowColors: Record<number, string>;
-  voicePeakRows: Record<number, number>;
-}
-
-/** Compute smoothed level and bar height for a voice */
-export function computeVoiceBar(state: VoiceVuState, voice: number): number {
-  const level = getVoiceLevel(voice);
-  const actualLevel = level ? Math.max(level.left, level.right) : 0;
-  state.smoothedLevel = Math.max(
-    actualLevel,
-    state.smoothedLevel * VU_DECAY_RATE,
-  );
-  // Below noise floor = dead silence (0 bar height)
-  if (state.smoothedLevel < VU_NOISE_FLOOR) return 0;
-  const scaled = Math.pow(
-    state.smoothedLevel * VU_AMPLIFICATION,
-    VU_POWER_CURVE,
-  );
-  return Math.min(ICON_ROWS, Math.max(0, scaled * ICON_ROWS));
-}
-
-/** Compute VU brightness for a single LED */
-export function computeVuLedBrightness(
-  row: number,
-  barHeight: number,
-  peakRow: number,
-): number {
-  const litThreshold = ICON_ROWS - barHeight;
-  const isLit = row >= litThreshold;
-  const isPeak = row === peakRow && peakRow >= 0;
-
-  if (isPeak) return VU_PEAK_BRIGHTNESS;
-  if (isLit) {
-    const posInBar =
-      barHeight > 0 ? (ICON_ROWS - 1 - row) / (ICON_ROWS - 1) : 0;
-    return (
-      VU_LIT_MIN_BRIGHTNESS +
-      posInBar * (VU_LIT_MAX_BRIGHTNESS - VU_LIT_MIN_BRIGHTNESS)
-    );
-  }
-  return VU_UNLIT_BRIGHTNESS;
-}
-
-export function decayPeak(state: VoiceVuState): void {
-  if (state.peakHoldCounter > 0) {
-    state.peakHoldCounter--;
-    return;
-  }
-  state.peakDropCounter++;
-  if (state.peakDropCounter >= VU_PEAK_DROP_INTERVAL) {
-    state.peakDropCounter = 0;
-    state.peakRow++;
-    if (state.peakRow >= ICON_ROWS) {
-      state.peakRow = -1;
-    }
-  }
-}
-
-export function getVuState(
-  states: Record<number, VoiceVuState>,
-  voice: number,
-): VoiceVuState {
-  if (!states[voice]) {
-    states[voice] = initVuState();
-  }
-  return states[voice];
-}
-
-export function initVuState(): VoiceVuState {
-  return {
-    peakDropCounter: 0,
-    peakHoldCounter: 0,
-    peakRow: -1,
-    smoothedLevel: 0,
-  };
-}
-
-export function updateCrossFade(
-  cf: CrossFadeState,
-  vuActive: boolean,
-  rawTime: number,
-  vuStates: Record<number, VoiceVuState>,
-): void {
-  if (vuActive) {
-    cf.wasVuActive = true;
-    cf.vuInactiveSince = null;
-    cf.ambientFade = 0;
-    return;
-  }
-
-  if (!cf.wasVuActive) {
-    cf.ambientFade = 1;
-    return;
-  }
-
-  const allDecayed = Object.values(vuStates).every(
-    (s) => s.smoothedLevel < VU_SILENCE_THRESHOLD,
-  );
-  if (!allDecayed) return;
-
-  cf.vuInactiveSince ??= rawTime;
-  const elapsed = (rawTime - cf.vuInactiveSince) * 1000;
-  cf.ambientFade = Math.min(1, elapsed / VU_CROSSFADE_MS);
-  if (cf.ambientFade >= 1) {
-    cf.wasVuActive = false;
-    cf.vuInactiveSince = null;
-  }
-}
-
-/** Update peak hold/drop state and return current peak row */
-export function updatePeakState(state: VoiceVuState, barHeight: number): void {
-  const currentTopRow = ICON_ROWS - Math.ceil(barHeight);
-
-  if (barHeight > 0 && currentTopRow <= state.peakRow) {
-    state.peakRow = currentTopRow;
-    state.peakHoldCounter = VU_PEAK_HOLD_FRAMES;
-    state.peakDropCounter = 0;
-    return;
-  }
-
-  if (state.peakRow >= 0) {
-    decayPeak(state);
-    return;
-  }
-
-  if (barHeight > 0) {
-    state.peakRow = currentTopRow;
-    state.peakHoldCounter = VU_PEAK_HOLD_FRAMES;
-    state.peakDropCounter = 0;
-  }
-}
-
+/**
+ * Animation loop for the LED icon. Owns the rAF driver, time/hover speed
+ * handling, and the pointer/ripple interaction API; per-frame VU state comes
+ * from vuMeterState and per-LED painting from ledRenderers.
+ */
 export function useLedVisualization(
   isHoveredRef: React.RefObject<boolean>,
 ): UseLedVisualizationReturn {
@@ -371,106 +185,4 @@ export function useLedVisualization(
   }, []);
 
   return { addRipple, clearMousePosition, ledRefs, setMousePosition };
-}
-
-function addRippleBoost(
-  col: number,
-  row: number,
-  rawTime: number,
-  ripples: Ripple[],
-): number {
-  if (ripples.length === 0) return 0;
-  return computeRippleBoost(
-    col,
-    row,
-    rawTime,
-    ripples,
-    ICON_RIPPLE_SPEED,
-    ICON_RIPPLE_WIDTH,
-    ICON_RIPPLE_DURATION,
-  );
-}
-
-function computeAmbientBrightness(
-  col: number,
-  row: number,
-  time: number,
-): number {
-  const config = ICON_ROW_LFO_CONFIGS[row];
-  return (
-    ICON_BASE_MIN +
-    computeBaseLfo(col, row, time, config, ICON_BASE_SCALE) * ICON_BASE_SCALE
-  );
-}
-
-function renderAmbientLed(options: AmbientLedOptions): void {
-  const { col, el, glowColor, isHovered, mouse, rawTime, ripples, row, time } =
-    options;
-  el.style.backgroundColor = "";
-  let brightness = computeAmbientBrightness(col, row, time);
-
-  if (isHovered && mouse) {
-    brightness += computeProximityBoost(
-      col,
-      row,
-      mouse.col,
-      mouse.row,
-      ICON_PROXIMITY_RADIUS_SQ,
-      ICON_PROXIMITY_INTENSITY,
-    );
-  }
-
-  brightness += addRippleBoost(col, row, rawTime, ripples);
-  applyLedStyle(el, brightness, glowColor);
-}
-
-function renderVuLed(options: VuLedOptions): void {
-  const {
-    ambientMix,
-    col,
-    defaultGlowColor,
-    el,
-    rawTime,
-    ripples,
-    row,
-    time,
-    voiceBarHeights,
-    voiceGlowColors,
-    voicePeakRows,
-  } = options;
-  const voice = colToVoice.get(col);
-  const isBorder = borderColSet.has(col);
-
-  let brightness: number;
-  let glowColor: string;
-
-  if (isBorder) {
-    brightness = VU_BORDER_BRIGHTNESS;
-    glowColor = defaultGlowColor;
-    el.style.backgroundColor = "";
-  } else if (voice) {
-    brightness = computeVuLedBrightness(
-      row,
-      voiceBarHeights[voice],
-      voicePeakRows[voice],
-    );
-    glowColor = voiceGlowColors[voice] ?? defaultGlowColor;
-    const alpha = brightness > VU_UNLIT_BRIGHTNESS ? brightness : 0.05;
-    el.style.backgroundColor = `rgba(${glowColor}, ${alpha.toFixed(2)})`;
-  } else {
-    brightness = VU_UNLIT_BRIGHTNESS;
-    glowColor = defaultGlowColor;
-    el.style.backgroundColor = "";
-  }
-
-  if (ambientMix > 0) {
-    const ambientBrightness = computeAmbientBrightness(col, row, time);
-    brightness = brightness * (1 - ambientMix) + ambientBrightness * ambientMix;
-    if (ambientMix > 0.5) {
-      el.style.backgroundColor = "";
-    }
-  }
-
-  brightness += addRippleBoost(col, row, rawTime, ripples);
-  applyLedStyle(el, brightness, glowColor);
 }
